Update config.example.js
Adds two configuration values in case the end user wants to vote for a server instead of a bot. The serverID is the guildID, while serverMode is a boolean flag that represents the voting mode. If true, the script votes for a server. If false, it reverts to default behavior and votes for a bot.